another attempt to fix global work size

+++ GPU/vector_add/matrix_mult.cpp
@@ -278,7 +278,7 @@ clGetPlatformIDs(1, &platform, NULL);
     status = clSetKernelArg(kernel, argi++, sizeof(cl_mem), &output_buf);
     checkError(status, "Failed to set argument 4");
 
-    const size_t[2] global_work_size = {N, N};
+    const unsigned int [2] global_work_size = {N, N};
     status = clEnqueueNDRangeKernel(queue, kernel, 2, NULL,
         &global_work_size, NULL, 2, write_event, &kernel_event);
     checkError(status, "Failed to launch kernel");
@@ -297,7 +297,7 @@ clGetPlatformIDs(1, &platform, NULL);
         pass = false;
       }
     }
-	
+
 
     //Define timing variables
     cl_ulong startgpu, endgpu, diffgpu;
